jpg2printablePDF/requirements.txt
Pillow
fpdf
pytest
